requirements_google.txt
python-telegram-bot
mistralai
PyMuPDF
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
